[feat] activity_main -> fragment_profile 옮김
