+++ app/src/main/res/layout/activity_main.xml
@@ -3,134 +3,30 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <data>
-
-        <variable
-            name="viewModel"
-            type="org.sopt.android_hyorim_30th.ui.main.MainViewModel" />
-    </data>
-
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:context=".ui.main.MainActivity">
 
-        <ScrollView
-            android:id="@+id/nestedScrollView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <androidx.constraintlayout.widget.ConstraintLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content">
-
-                <ImageView
-                    android:id="@+id/profile_photo"
-                    android:layout_width="0dp"
-                    android:layout_height="0dp"
-                    android:layout_marginHorizontal="130dp"
-                    android:layout_marginTop="20dp"
-                    android:scaleType="fitXY"
-                    app:imageUri="@{viewModel.profileImage}"
-                    app:layout_constraintDimensionRatio="1:1"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
-
-                <TextView
-                    android:id="@+id/tv_name"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="80dp"
-                    android:layout_marginTop="20dp"
-                    android:text="@{`이름 : ` + viewModel.userName}"
-                    android:textColor="@color/black"
-                    android:textStyle="bold"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@+id/profile_photo"
-                    tools:text="이름" />
-
-                <TextView
-                    android:id="@+id/tv_age"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="10dp"
-                    android:text="@{`나이    ` +  String.valueOf(viewModel.userAge)}"
-                    android:textColor="@color/black"
-                    android:textStyle="bold"
-                    app:layout_constraintStart_toStartOf="@id/tv_name"
-                    app:layout_constraintTop_toBottomOf="@+id/tv_name"
-                    tools:text="나이" />
-
-                <TextView
-                    android:id="@+id/tv_mbti"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="10dp"
-                    android:text="@{`MBTI    ` + viewModel.userMBTI}"
-                    android:textColor="@color/black"
-                    android:textStyle="bold"
-                    app:layout_constraintStart_toStartOf="@id/tv_age"
-                    app:layout_constraintTop_toBottomOf="@+id/tv_age"
-                    tools:text="MBTI" />
-
-                <TextView
-                    android:id="@+id/tv_introduction"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="10dp"
-                    android:layout_marginBottom="10dp"
-                    android:text="@{viewModel.userIntroduction}"
-                    android:textAlignment="center"
-                    android:textColor="@color/black"
-                    android:textStyle="bold"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintStart_toStartOf="@id/tv_mbti"
-                    app:layout_constraintTop_toBottomOf="@+id/tv_mbti"
-                    tools:text="자기소개" />
-
-            </androidx.constraintlayout.widget.ConstraintLayout>
-
-        </ScrollView>
-
-        <Button
-            android:id="@+id/btn_follower"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="10dp"
-            android:layout_marginEnd="20dp"
-            android:backgroundTint="@color/purple_200"
-            android:onClick="@{()->viewModel.onClickFollowerBtn()}"
-            android:paddingHorizontal="30dp"
-            android:paddingVertical="15dp"
-            android:text="@string/follower_list"
-            app:layout_constraintEnd_toStartOf="@id/btn_repository"
-            app:layout_constraintHorizontal_chainStyle="packed"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/nestedScrollView" />
-
-        <Button
-            android:id="@+id/btn_repository"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:backgroundTint="@color/purple_500"
-            android:onClick="@{()->viewModel.onClickRepositoryBtn()}"
-            android:paddingHorizontal="30dp"
-            android:paddingVertical="15dp"
-            android:text="@string/repository_list"
-            app:layout_constraintBottom_toBottomOf="@id/btn_follower"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toEndOf="@id/btn_follower"
-            app:layout_constraintTop_toTopOf="@id/btn_follower" />
-
-        <androidx.fragment.app.FragmentContainerView
-            android:id="@+id/container"
+        <androidx.viewpager2.widget.ViewPager2
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            android:layout_marginTop="10dp"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/btn_follower" />
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <!--        <com.google.android.material.bottomnavigation.BottomNavigationView-->
+        <!--            android:id="@+id/bnv_sample"-->
+        <!--            android:layout_width="match_parent"-->
+        <!--            android:layout_height="wrap_content"-->
+        <!--            android:theme="@style/Widget.BottomNavigationView"-->
+        <!--            app:itemIconTint="@color/selector_bottom_navi"-->
+        <!--            app:itemRippleColor="#FFFFFF"-->
+        <!--            app:itemTextColor="@color/selector_bottom_navi"-->
+        <!--            app:layout_constraintBottom_toBottomOf="parent"-->
+        <!--            app:layout_constraintEnd_toEndOf="parent"-->
+        <!--            app:layout_constraintStart_toStartOf="parent"-->
+        <!--            app:menu="@menu/menu_sample" />-->
+
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
